nfs: Fix the return value of nfs4_select_rw_stateid
Patchwork-id: 79018
O-Subject: [RHEL6.6 PATCH 4/7] NFSv4: Fix the return value of nfs4_select_rw_stateid
Bugzilla: 1075123
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

In commit 5521abfdcf4d6 (NFSv4: Resend the READ/WRITE RPC call
if a stateid change causes an error), we overloaded the return value of
nfs4_select_rw_stateid() to cause it to return -EWOULDBLOCK if an RPC
call is outstanding that would cause the NFSv4 lock or open stateid
to change.
That is all redundant when we actually copy the stateid used in the
read/write RPC call that failed, and check that against the current
stateid. It is doubly so, when we consider that in the NFSv4.1 case,
we also set the stateid's seqid to the special value '0', which means
'match the current valid stateid'.

Reported-by: Andy Adamson <andros@netapp.com>
Link: http://lkml.kernel.org/r/1393954269-3974-1-git-send-email-andros@netapp.com
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream Id: 927864cd92aaad1d6285e3bb16e503caf3d6e27e
Bugzilla: 1075123

Signed-off-by: Rafael Aquini <aquini@redhat.com>