tools: Let O= makes handle relative paths
Patchwork-id: 79081
O-Subject: [PATCH RHEL6.6 BZ1070896 03/89] perf tools: Let O= makes handle relative paths
Bugzilla: 1070896
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1070896
https://bugzilla.redhat.com/show_bug.cgi?id=1070896

upstream
========
commit c883122acc0d97648d8b8f4726709017674e4420
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Mon Aug 13 10:23:02 2012 -0400

description
===========
When I did a compile of perf using a relative path for the output
directory, the build failed when it tried to compile libtraceevent. This
is because it continues to use the same relative path when the new
working directory is in a different path.

    SUBDIR ../lib/traceevent/
/bin/sh: line 0: cd: ../../../nobackup/perf/: No such file or directory
Makefile:74: *** output directory "../../../nobackup/perf/" does not exist.  Stop.
make: *** [../../../nobackup/perf///libtraceevent.a] Error 2

Make the path used an absolute path when building perf with O=.

Boris:

Teach Makefile to check whether the supplied O= directory exists and
bail out if not. Reportedly, kernel dudes are idiots and need to be
guarded so as not to shoot themselves in the foot when playing in the
sandbox.

Signed-off-by: Rafael Aquini <aquini@redhat.com>