tools: Honour the O= flag when tool build called from a higher Makefile
Patchwork-id: 79084
O-Subject: [PATCH RHEL6.6 BZ1070896 05/89] tools: Honour the O= flag when tool build called from a higher Makefile
Bugzilla: 1070896
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1070896
https://bugzilla.redhat.com/show_bug.cgi?id=1070896

upstream
========
commit bf35182ffcd00d8b36d56210ffdac110e5624d7d
Author: David Howells <dhowells@redhat.com>
Date: Mon Nov 5 21:02:08 2012 +0000

description
===========
Honour the O= flag that was passed to a higher level Makefile and then passed
down as part of a tool build.

To make this work, the top-level Makefile passes the original O= flag and
subdir=tools to the tools/Makefile, and that in turn passes
subdir=$(O)/$(subdir)/foodir when building tool foo in directory
$(O)/$(subdir)/foodir (where the intervening slashes aren't added if an
element is missing).

For example, take perf.  This is found in tools/perf/.  Assume we're building
into directory ~/zebra/, so we pass O=~/zebra to make.  Dependening on where
we run the build from, we see:

	make run in dir		$(OUTPUT) dir
	=======================	==================
	linux			~/zebra/tools/perf/
	linux/tools		~/zebra/perf/
	linux/tools/perf	~/zebra/

and if O= is not set, we get:

	make run in dir		$(OUTPUT) dir
	=======================	==================
	linux			linux/tools/perf/
	linux/tools		linux/tools/perf/
	linux/tools/perf	linux/tools/perf/

The output directories are created by the descend function if they don't
already exist.

Signed-off-by: Rafael Aquini <aquini@redhat.com>