tools: Correct Makefile.include
Patchwork-id: 79085
O-Subject: [PATCH RHEL6.6 BZ1070896 08/89] perf tools: Correct Makefile.include
Bugzilla: 1070896
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1070896
https://bugzilla.redhat.com/show_bug.cgi?id=1070896

upstream
========
commit 9e4a66482e9a96405eb9ee7f7bf28c9799ca8670
Author: Borislav Petkov <bp@suse.de>
Date: Wed Feb 20 16:32:29 2013 +0100

description
===========
It looks at O= and adjusts the $(OUTPUT) variable based on what the
output directory will be. However, when O is defined but empty, it
wrongly becomes the user's $HOME dir which is not what we want. So check
it is not empty before working with it further.

Signed-off-by: Rafael Aquini <aquini@redhat.com>