annotate: Add a comment on the symbol__parse_objdump_line()
Patchwork-id: 79100
O-Subject: [PATCH RHEL6.6 BZ1070896 26/89] perf annotate: Add a comment on the symbol__parse_objdump_line()
Bugzilla: 1070896
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1070896
https://bugzilla.redhat.com/show_bug.cgi?id=1070896

upstream
========
commit 3aec150af3de6c00570bdacf45bf5a999ab9cf1d
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Mar 5 14:53:22 2013 +0900

description
===========
The symbol__parse_objdump_line() parses result of the objdump run but
it's hard to follow if one doesn't know the output format of the
objdump.  Add a head comment on the function to help her.

Signed-off-by: Rafael Aquini <aquini@redhat.com>