stat: Introduce --repeat forever
Patchwork-id: 79130
O-Subject: [PATCH RHEL6.6 BZ1070896 54/89] perf stat: Introduce --repeat forever
Bugzilla: 1070896
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1070896
https://bugzilla.redhat.com/show_bug.cgi?id=1070896

upstream
========
commit a7e191c376fad084d9f3c7ac89a1f7c47462ebc8
Author: Frederik Deweerdt <frederik.deweerdt@xprog.eu>
Date: Fri Mar 1 13:02:27 2013 -0500

description
===========
The following patch causes 'perf stat --repeat 0' to be interpreted as
'forever', displaying the stats for every run.

We act as if a single run was asked, and reset the stats in each
iteration. In this mode SIGINT is passed to perf to be able to stop the
loop with Ctrl+C.

Signed-off-by: Rafael Aquini <aquini@redhat.com>