Add PERF_RECORD_MISC_MMAP_DATA to RECORD_MMAP
Patchwork-id: 79178
O-Subject: [PATCHv2 RHEL6.6 BZ1059205 16/20] perf: Add PERF_RECORD_MISC_MMAP_DATA to RECORD_MMAP
Bugzilla: 1059205
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1059205
https://bugzilla.redhat.com/show_bug.cgi?id=1059205

upstream
========
commit 2fe85427e3bf65d791700d065132772fc26e4d75
Author: Stephane Eranian <eranian@google.com>
Date:   Thu Jan 24 16:10:39 2013 +0100

description
===========
Type of mapping was lost and made it hard for a tool
to distinguish code vs. data mmaps. Perf has the ability
to distinguish the two.

Use a bit in the header->misc bitmask to keep track of
the mmap type. If PERF_RECORD_MISC_MMAP_DATA is set then
the mapping is not executable (!VM_EXEC). If not set, then
the mapping is executable.

Signed-off-by: Rafael Aquini <aquini@redhat.com>