gfs2: avoid locking while unfreezing fs
Patchwork-id: 79286
O-Subject: [RHEL 6.6 PATCH BZ#1056625] GFS2: avoid locking while unfreezing fs
Bugzilla: 1056625
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

If an operation tries to grab the filesystem root directory glock
exclusively while the filesystem is frozen, the filesystem will hang
when it attempts to unfreeze.  This is because fsfreeze needs to grab a
shared glock on the filesystem root directory before it issues the ioctl
to do the unfreeze.

In order to allow the unfreeze, this patch makes gfs2 grab a shared lock
on the filesystem root directory during the freeze, and hold it until it
unfreezes the filesystem.  The functions which need to grab a shared
lock in order to allow the unfreeze ioctl to be issued now use the
locked grabbed by the freeze code instead.

The freeze and unfreeze code take care to make sure that this shared
lock will not be dropped while another process is using it.

This patch is part of a much larger patch being submitted upstream to
fix freezing issues, and redo some gfs2 locking.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>