gfs2: Don't use gfs2_change_nlink in link syscall
Patchwork-id: 79371
O-Subject: [RHEL6.6 PATCH 1/3] GFS2: Don't use gfs2_change_nlink in link syscall
Bugzilla: 1092083
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

There are three users of gfs2_change_nlink which add to the link
count. Two of these are about to be removed in later patches, so
this means that there will no callers, when that happens allowing
removal of that function, also in a later patch.

rhbz#1092083

Signed-off-by: Rafael Aquini <aquini@redhat.com>