gfs2: When adding a new dir entry, inc link count if it is a subdir
Patchwork-id: 79372
O-Subject: [RHEL6.6 PATCH 3/3] GFS2: When adding a new dir entry, inc link count if it is a subdir
Bugzilla: 1092083
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

This adds an increment of the link count when we add a new directory
entry, if that entry is itself a directory. This means that we no
longer need separate code to perform this operation.

Now that both adding and removing directory entries automatically
update the parent directory's link count if required, that makes
the code shorter and simpler than before.

rhbz#1092083

Signed-off-by: Rafael Aquini <aquini@redhat.com>