btrfs: don't oops on mount failure due to unsupported features
Patchwork-id: 79374
O-Subject: [RHEL6 PATCH] btrfs: don't oops on mount failure due to unsupported features
Bugzilla: 889471
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This is for bug
889471 - [Btrfs] BUG: unable to handle kernel NULL pointer dereference at (null) btrfs_get_sb should return error when open_ctree failed

HP reported it too; if you have a disk lying around that was formatted
with newer btrfs-progs (say, from RHEL7) and mount it under RHEL6,
the box will panic, because we return success when get_default_root
has actually failed.

This isn't upstream, exactly; other things have changed around it,
but this captures the spirit of the (still incomplete) error
handling upstream I think... it's not beautiful, but it solves
the problem in my testing.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>