allow a partially recovered device to be hot-added to an array
Patchwork-id: 79638
O-Subject: [RHEL6.6 PATCH 46/58] md: allow a partially recovered device to be hot-added to an array.
Bugzilla: 1030619
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: NeilBrown <neilb@suse.de>

When adding a new device into an array it is normally important to
clear any stale data from ->recovery_offset else the new device may
not be recovered properly.

However when re-adding a device which is known to be nearly in-sync,
this is not needed and can be detrimental.  The (bitmap-based)
resync will still happen, and further recovery is only needed from
where-ever it was already up to.

So if save_raid_disk is set, signifying a re-add, don't clear
->recovery_offset.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit 7eb418851f3278de67126ea0c427641ab4792c57)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>