xfs: fix directory readahead offset off-by-one
Patchwork-id: 79661
O-Subject: [PATCH RHEL6.6] xfs: fix directory readahead offset off-by-one
Bugzilla: 1093820
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Dave Chinner <dchinner@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This is for:

Bug 1093820 - possible read() corruption XFS on kernel 2.6.32-358.41.1 (redhat storage)

and it's a backport of this patch from for-next in the upstream xfs tree:

http://oss.sgi.com/cgi-bin/gitweb.cgi?p=xfs/xfs.git;a=commitdiff;h=8cfcc3e565bf15870efe801368a25ca98092e6e7

It's been tested with the customer-supplied metadump, and it solves
the problem.

Thanks,
-Eric

Directory readahead can throw loud scary but harmless warnings
when multiblock directories are in use a specific pattern of
discontiguous blocks are found in the directory. That is, if a hole
follows a discontiguous block, it will throw a warning like:

XFS (dm-1): xfs_da_do_buf: bno 637 dir: inode 34363923462
XFS (dm-1): [00] br_startoff 637 br_startblock 1917954575 br_blockcount 1 br_state 0
XFS (dm-1): [01] br_startoff 638 br_startblock -2 br_blockcount 1 br_state 0

And dump a stack trace.

This is because the readahead offset increment loop does a double
increment of the block index - it does an increment for the loop
iteration as well as increase the loop counter by the number of
blocks in the extent. As a result, the readahead offset does not get
incremented correctly for discontiguous blocks and hence can ask for
readahead of a directory block from an offset part way through a
directory block.  If that directory block is followed by a hole, it
will trigger a mapping warning like the above.

The bad readahead will be ignored, though, because the main
directory block read loop uses the correct mapping offsets rather
than the readahead offset and so will ignore the bad readahead
altogether.

Fix the warning by ensuring that the readahead offset is correctly
incremented.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>