ext4: add mmp fs freeze protection
Patchwork-id: 79786
O-Subject: [RHEL6.6 PATCH 6/6 V2] ext4: add mmp fs freeze protection
Bugzilla: 1012412
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

This is for
Bug 1012412 - [RFE] EXT4: Prevent mounting SAN device on two nodes

This is a small part of the upstream commit
8e8ad8a5 ext4: Convert to new freezing mechanism
which is already in RHEL6, but at the time it was merged,
mmp.c did not exist, so fix that up now.

V2: Pass in the *sb that we need.  Whoops.
Signed-off-by: Rafael Aquini <aquini@redhat.com>