x86: fix kernel crash with PEBS/BTS after suspend/resume
Patchwork-id: 80026
O-Subject: [PATCH RHEL6.6 BZ829875 04/43] perf, x86: fix kernel crash with PEBS/BTS after suspend/resume
Bugzilla: 829875
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 829875
https://bugzilla.redhat.com/show_bug.cgi?id=829875

upstream
========
commit 1d9d8639c063caf6efc2447f5f26aa637f844ff6
Author: Stephane Eranian <eranian@google.com>
Date: Fri Mar 15 14:26:07 2013 +0100

description
===========
This patch fixes a kernel crash when using precise sampling (PEBS)
after a suspend/resume. Turns out the CPU notifier code is not invoked
on CPU0 (BP). Therefore, the DS_AREA (used by PEBS) is not restored properly
by the kernel and keeps it power-on/resume value of 0 causing any PEBS
measurement to crash when running on CPU0.

The workaround is to add a hook in the actual resume code to restore
the DS Area MSR value. It is invoked for all CPUS. So for all but CPU0,
the DS_AREA will be restored twice but this is harmless.

Signed-off-by: Rafael Aquini <aquini@redhat.com>