x86: Fix uninitialized pt_regs in intel_pmu_drain_bts_buffer()
Patchwork-id: 80027
O-Subject: [PATCH RHEL6.6 BZ829875 06/43] perf/x86: Fix uninitialized pt_regs in intel_pmu_drain_bts_buffer()
Bugzilla: 829875
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 829875
https://bugzilla.redhat.com/show_bug.cgi?id=829875

upstream
========
commit 0e48026ae7abf871e51eaa9183c81ab5bef4c267
Author: Stephane Eranian <eranian@google.com>
Date: Tue Mar 19 16:10:38 2013 +0100

description
===========
This patch fixes an uninitialized pt_regs struct in drain BTS
function. The pt_regs struct is propagated all the way to the
code_get_segment() function from perf_instruction_pointer()
and may get garbage.

We cannot simply inherit the actual pt_regs from the interrupt
because BTS must be flushed on context-switch or when the
associated event is disabled. And there we do not have a pt_regs
handy.

Setting pt_regs to all zeroes may not be the best option but it
is not clear what else to do given where the drain_bts_buffer()
is called from.

In V2, we move the memset() later in the code to avoid doing it
when we end up returning early without doing the actual BTS
processing. Also dropped the reg.val initialization because it
is redundant with the memset() as suggested by PeterZ.

Signed-off-by: Rafael Aquini <aquini@redhat.com>