x86: Add mem-loads/stores support for Haswell
Patchwork-id: 80034
O-Subject: [PATCH RHEL6.6 BZ829875 13/43] perf/x86/intel: Add mem-loads/stores support for Haswell
Bugzilla: 829875
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 829875
https://bugzilla.redhat.com/show_bug.cgi?id=829875

upstream
========
commit f9134f36aed59ab55c0ab1a4618dd455f15aef5f
Author: Andi Kleen <ak@linux.intel.com>
Date: Mon Jun 17 17:36:52 2013 -0700

description
===========
mem-loads is basically the same as Sandy Bridge,
but we use a separate string for changes later.

Haswell doesn't support the full precise store mode,
so we emulate it using the "DataLA" facility.
This allows to do everything, but for data sources we
can only detect L1 hit or not.

There is no explicit enable bit anymore, so we have
to tie it to a perf internal only flag.

The address is supported for all memory related PEBS
events with DataLA. Instead of only logging for the
load and store events we allow logging it for all
(it will be simply 0 if the current event does not
support it)

Signed-off-by: Rafael Aquini <aquini@redhat.com>