x86: Report TSX transaction abort cost as weight
Patchwork-id: 80039
O-Subject: [PATCH RHEL6.6 BZ829875 17/43] perf/x86: Report TSX transaction abort cost as weight
Bugzilla: 829875
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 829875
https://bugzilla.redhat.com/show_bug.cgi?id=829875

upstream
========
commit 748e86aa90edfddfa6016f1cf383ff5bc6aada91
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Sep 5 20:37:39 2013 -0700

description
===========
Use the existing weight reporting facility to report the transaction
abort cost, that is the number of cycles wasted in aborts.
Haswell reports this in the PEBS record.

This was in fact the original user for weight.

This is a very useful sort key to concentrate on the most
costly aborts and a good metric for TSX tuning.

Signed-off-by: Rafael Aquini <aquini@redhat.com>