hists: Set rules hint for the hist browser
Patchwork-id: 80689
O-Subject: [PATCH RHEL6.6 BZ1101039 045/233] perf gtk/hists: Set rules hint for the hist browser
Bugzilla: 1101039
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1101039
https://bugzilla.redhat.com/show_bug.cgi?id=1101039

upstream
========
commit 9d58d2f66c92fe47dd395947a3d51b9ace7dcc92
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Jun 4 18:22:17 2013 +0900

description
===========
The 'rules' means that every second line of the tree view has a shaded
background, which makes it easier to see which cell belongs to which
row in the tree view.  It can be useful for a tree view that has a lot
of rows.