diff: Switching the base hists to be pairs head
Patchwork-id: 80691
O-Subject: [PATCH RHEL6.6 BZ1101039 056/233] perf diff: Switching the base hists to be pairs head
Bugzilla: 1101039
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1101039
https://bugzilla.redhat.com/show_bug.cgi?id=1101039

upstream
========
commit 9af303e22a317d1cc6f440e08f72428830708b37
Author: Jiri Olsa <jolsa@redhat.com>
Date: Sat Dec 1 21:15:40 2012 +0100

description
===========
Making the baseline hists to act as a pairs head.

So far we don't care which hists act as a pairs head, because we have
only 2 files to deal with and any of them is suitable to do the job.

But if we want to process more files, we need to pick up one hists to
act as pairs head, and the baseline hists is the most suitable.