annotate: Improve description of '?' hotkey
Patchwork-id: 80764
O-Subject: [PATCH RHEL6.6 BZ1101039 113/233] perf annotate browser: Improve description of '?' hotkey
Bugzilla: 1101039
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1101039
https://bugzilla.redhat.com/show_bug.cgi?id=1101039

upstream
========
commit fcd9fef9a64b84520852e0247d6796893e28864c
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Aug 7 15:55:48 2013 -0300

description
===========
The previous description: "Search previous string" is usually associated
with the 'N' following a '/string', the opposite of 'n', which is
'Search next string' in the direction established with '/' or '?'.

So change it to 'Search string backwards', to clarify that.

The 'N' hotkey remains to be implemented with the semantic described
above.