mpt2sas: when Async scanning is enabled then while scanning, devices are removed but their transport layer entries are not removed
Patchwork-id: 81535
O-Subject: [RHEL6.6 e-stor PATCH 07/11] mpt2sas: when Async scanning is enabled then while scanning, devices are removed but their transport layer entries are not removed
Bugzilla: 1059137
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

When Async scanning mode is enabled and device scanning is in progress then
devices should not be removed. But in actuality, devices are removed but
their transport layer entries are not removed. This causes error to add
the same device to the transport layer after host reset or diagnostic
reset.

So, in this patch, modified the code in such a way that device is not removed
when Async scanning mode is enabled and device scanning is in progress.

Signed-off-by: Sreekanth Reddy <Sreekanth.Reddy@lsi.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
(cherry picked from commit bc6d4c3b78d1320606f99905e2a9fa86b246b449)

Signed-off-by: Rafael Aquini <aquini@redhat.com>