gfs2: Ensure workqueue is scheduled after noexp request
Patchwork-id: 81565
O-Subject: [RHEL6.6 PATCH][bz#1101740] GFS2: Ensure workqueue is scheduled after noexp request
Bugzilla: 1101740
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is a back-port of the following upstream patch:

http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2/glock.c?id=01b172b7b10146cf5f02604047bee065cfb49946

The difference between the upstream and RHEL6.6 patch is that the glock
reference count is atomic in RHEL6.6 and it isn't atomic upstream.

A kernel build that includes this patch was done here:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7517897

With this patch, QE's "revolver" GFS2 recovery test has passed 50
iterations.

Patch description:

This patch closes a small timing window whereby a request to hold the
transaction glock can get stuck. The problem is that after the DLM has
granted the lock, it can get into a state whereby it doesn't transition
the glock to a held state, due to not having requeued the glock state
machine to finish the transition.

rhbz#1101740

Signed-off-by: Rafael Aquini <aquini@redhat.com>