sfc: Ensure MCDI buffers, but not lengths, are dword aligned
Patchwork-id: 81663
O-Subject: [RHEL6.6 PATCH BZ 1057527 012/179] sfc: Ensure MCDI buffers, but not lengths, are dword aligned
Bugzilla: 1057527
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

From: Ben Hutchings <bhutchings@solarflare.com>

We currently require that MCDI request and response lengths are
multiples of 4 bytes, because we will copy dwords in and out of shared
memory and we want to be sure we won't read or write out of bounds.
But all we really need to know is that there is sufficient padding for
that.  Also, we should ensure that buffers are dword-aligned, as on
some architectures misaligned access will result in data corruption or
a crash.

Change the buffer type to array-of-efx_dword_t and remove the
requirement that the lengths are multiples of 4.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>
(cherry picked from commit 9528b9219348e0a013f4b587958a8ba9c96d7e20)

Additional change: adjust our private (RHEL-specific) ioctl.c type
casting to denote the changed type.