cifs: Check if prefixpath starts with '\' in cifs_parse_mount_options
Patchwork-id: 83030
O-Subject: [RHEL6 PATCH] [bz1104268]cifs: Check if prefixpath starts with '\' in cifs_parse_mount_options
Bugzilla: 1104268
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

cifs_mount() calls convert_delimiter() to change all instances of '/' in
the prefixpath to '\'. When dealing with DFS shares, we call
cifs_parse_mount_options() again with the now modified prefix path to
create a new volume info structure for the node we are being redirected
to. Since we only check for a leading '/' here, we ignore the leading
'\' and end up prefixing '/' which then gets converted to '\' ending up
with a prefix path beginning with "\\".

Fix this by also checking for a leading '\' for the prefixpath in
cifs_parse_mount_options()

This problem has been addressed upstream with commit
839db3d10a5ba792d6533b8bb3380f52ac877344
This patch however modified the newer versions of
cifs_parse_mount_options and changes some functionality which we would
like to avoid for RHEL 6.

A test kernel containing this patch has been tested successfully both
internally by the customer.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>