bonding: look for bridge IPs in arp monitoring
Patchwork-id: 83075
O-Subject: [PATCH v4 RHEL6] bonding: look for bridge IPs in arp monitoring
Bugzilla: 704190
RH-Acked-by: Vlad Yasevich <vyasevic@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>

RHBZ#704190

Upstream: RHEL-only

Description:

There's a long-standing bug in bonding for arp monitoring configuration
when the bridge is used - basically, bonding is not aware of any bridge
that is on top of itself that might have the desired IP address from which
the arp requests should be sent.

slaves -> bond0 -> bridge0 [IP]

That's a classical VM-related configuration which is used widely.

Upstream had the same issue and was fixed by introducing a lot of changes:

1) rx_handler
2) netdev_upper
3) netdev_adjacent
4) netdev_for_each_upper

On the first approach it would take at least 40 patches to backport all the
needed infrastructure, so I've decided to take the RHEL-only approach for
this change - considering that we already have quite a lot of RHEL-specific
code for bond+vlan interaction, and even if we, some day, decide to
backport everything to RHEL6 this change can be easily reverted.

The main problem is that, in both sending ("from which src IP to send") and
receiving ("is that our dst IP?") part of bonding arp monitor we are not
aware of any bridge that contains the bonding device as a port (or the vlan
device on top of bonding as a port).

Thankfully, we have a bond_confirm_addr() which takes a device parameter
(either bond->dev or vlan->dev) and tries to find out if there is the
requested IP assigned.

So, to fix this we just verify (in case we didn't find the address) if the
device provided is a bridge port and, if yes, search for the requested
address at the bridge device too.

Also, before going through bond's vlans, verify if the route found is for
bond->dev OR (if bonding is a direct bridge port) bond->br_port->br->dev.

Signed-off-by: Rafael Aquini <aquini@redhat.com>