fs: preserve kABI
Patchwork-id: 83189
O-Subject: [RHEL6.6 Patch V2 003/434] fs: preserve kABI
Bugzilla: 1029861 1030158 1043993 1051207 1053497 1054375 1059083 1059086 1059087 1059089 1062766 1078538 696003 807078 826288
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The commit 1a7bd2265fc removed the export of get_unused_fd and replaced
it with get_unused_fd_flags and made get_unused_fd a macro that calls
the new function.  We can't do that as we have to preserve our kABI, so
export both calls instead.

This is needed because any code that used get_unused_fd_flags prior to
now would get a macro that evaluated to alloc_fd, and alloc_fd is not
exported from the kernel, so the module would end up with unknown
symbols after the compile was complete.

Signed-off-by: Doug Ledford <dledford@redhat.com>