mlx4_core: Fix checking order in MR table init
Patchwork-id: 83316
O-Subject: [RHEL6.6 Patch V2 122/434] net/mlx4_core: Fix checking order in MR table init
Bugzilla: 1029861 1030158 1043993 1051207 1053497 1054375 1059083 1059086 1059087 1059089 1062766 1078538 696003 807078 826288
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jack Morgenstein <jackm@dev.mellanox.co.il>

In procedure mlx4_init_mr_table(), slaves should do no processing,
but should return success. This initialization is hypervisor-only.

However, the check for num_mpts being a power-of-2 was performed
before the check to return immediately if the driver is for a slave.
This resulted in spurious failures.

The order of performing the checks is reversed, so that if the
driver is for a slave, no processing is done and success is returned.

Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Cherry picked from a30f1bc5c02

Signed-off-by: Doug Ledford <dledford@redhat.com>