mlx5: use uintptr_t casts to work around gcc warnings
Patchwork-id: 83393
O-Subject: [RHEL6.6 Patch V2 283/434] net/mlx5: use uintptr_t casts to work around gcc warnings
Bugzilla: 1029861 1030158 1043993 1051207 1053497 1054375 1059083 1059086 1059087 1059089 1062766 1078538 696003 807078 826288
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

In the mlx5 debugfs code, we have some u64 variables that are sometimes
filled with 64bit ints and sometimes with pointers.  This works fine on
64bit arches, but fails on 32bit arches because gcc's "error casting
from pointer to integer of a different size" doesn't distinguish the
case where the int is larger than the pointer and no data will be lost.
So, use the uintptr_t cast in between the pointer and u64 cast in order
to prevent the gcc errors on 32bit arches.

Signed-off-by: Doug Ledford <dledford@redhat.com>