hyperv: Enable sendbuf mechanism on the send path
Patchwork-id: 83987
O-Subject: [RHEL6.6 PATCH 21/23] hyperv: Enable sendbuf mechanism on the send path
Bugzilla: 1102261
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1102261

Notes:
- upstream commit b679ef73edc251f6d200a7dd2396e9fef9e36fc3 "hyperv: Add support
  for physically discontinuous receive buffer" was not backported since RHEL6
  does not have slow_virt_to_phys() support. So its follow-up:
  99d3016de4f2a29635f5382b0e9bd0e5f2151487 "hyperv: Change the receive buffer
  size for legacy hosts" was not backported either. We still use 2MB for receive
  buffer size which results the conflict.

We send packets using a copy-free mechanism (this is the Guest to Host transport
via VMBUS). While this is obviously optimal for large packets,
it may not be optimal for small packets. Hyper-V host supports
a second mechanism for sending packets that is "copy based". We implement that
mechanism in this patch.

In this version of the patch I have addressed a comment from David Miller.

With this patch (and all of the other offload and VRSS patches), we are now able
to almost saturate a 10G interface between Linux VMs on Hyper-V
on different hosts - close to  9 Gbps as measured via iperf.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Reviewed-by: Haiyang Zhang <haiyangz@microsoft.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit c25aaf814a63f9d9c4e45416f13d70ef0aa0be2e)

Conflicts:

	drivers/net/hyperv/hyperv_net.h
	drivers/net/hyperv/netvsc.c

Signed-off-by: Jason Wang <jasowang@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>