kabi: Adding symbol netdev_notice
Patchwork-id: 84212
O-Subject: [PATCH RHEL6.6 11/31] kabi: Adding symbol netdev_notice
Bugzilla: 1104850
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Michal Schmidt <mschmidt@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1104850
https://bugzilla.redhat.com/show_bug.cgi?id=1104850

Description:
============
BZ 1029841 introduced a backport of a conversion of dev_<level> and
netdev_<level> logging macros to proper functions. The advantage of
the functions is additional safety if code is recompiled to take
advantage of them.

If a 3rd party module is recompiled against RHEL6.6 headers, it may
start to depend on some of the newly added symbols, so we should
whitelist them.

Signed-off-by: Rafael Aquini <aquini@redhat.com>