gfs2: Only wait for demote when last holder is dequeued
Patchwork-id: 84298
O-Subject: [RHEL6.6 PATCH] [bz#1101740] GFS2: Only wait for demote when last holder is dequeued
Bugzilla: 1110421
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is for bz #1101740 - When using flock() a GFS2 filesystem hangs

It was tested on system gfs-i16c-02.mpc.lab.eng.bos.redhat.com using
a simple program (from comment #5 of the bz) that faithfully recreated
the problem. With the patch, the problem no longer recreates.

I submitted this patch upstream today.

Patch text:

Function gfs2_glock_dq_wait is supposed to dequeue a glock and then
wait for the lock to be demoted. The problem is, if this is a shared
lock, its demote will depend on the other holders, which means you
might end up waiting forever because the other process is blocked.
This problem is especially apparent when dealing with nested flocks.

rhbz#1101740

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>