nfs: Don't mark the data cache as invalid if it has been flushed
Patchwork-id: 85095
O-Subject: [RHEL6.6 PATCH 3/3] [bz1109365] NFS: Don't mark the data cache as invalid if it has been flushed
Bugzilla: 1109365
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

Now that we have functions such as nfs_write_pageuptodate() that use
the cache_validity flags to check if the data cache is valid or not,
it is a little more important to keep the flags in sync with the
state of the data cache.
In particular, we'd like to ensure that if the data cache is empty, we
don't start marking it as needing revalidation.

Reported-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
(cherry picked from commit 6edf96097b1df34fcd332f7ee4ba120fbc1d16c3)
Signed-off-by: Scott Mayhew <smayhew@redhat.com>

Conflicts:
	fs/nfs/inode.c

Signed-off-by: Rafael Aquini <aquini@redhat.com>