ehci: simplify remainder computations
Patchwork-id: 85175
O-Subject: [RHEL6 PATCH 4/7] USB: EHCI: simplify remainder computations
Bugzilla: 1092967
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1092967

commit bccbefaae050186bed3bcc74b1fd1a9b8c6710b2
Author: Alan Stern <stern@rowland.harvard.edu>
Date:   Wed Jul 14 11:03:36 2010 -0400

USB: EHCI: simplify remainder computations

This patch (as1406) adds a micro-optimization to ehci-hcd's scheduling
code.  Instead of computing remainders with respect to the schedule
length, use bitwise-and (which is quicker).  We know that the schedule
length will always be a power of two, but the compiler doesn't have
this information.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
CC: David Brownell <david-b@pacbell.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Rafael Aquini <aquini@redhat.com>