apic: Remove noisy zero-mask warning from default_send_IPI_mask_logical()
Patchwork-id: 85278
O-Subject: [RHEL6.6 PATCH BZ 745124] x86/apic: Remove noisy zero-mask warning from default_send_IPI_mask_logical()
Bugzilla: 745124
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=745124

Also seen in reports below as well as 629689 and 695731.  Recently seen in
nightly RHEL6 testing by Paul Bunyan.

commit e3f0f36ddf1b2743a0d4ea312996536a9c37e1c7
Author: Dave Jones <davej@redhat.com>
Date:   Fri Jan 18 12:58:47 2013 -0500

    x86/apic: Remove noisy zero-mask warning from default_send_IPI_mask_logical()

    Since circa 3.5, we've had dozens of reports of people hitting
    this warning. Forwarded reports have been met with silence, so
    just remove the warning if no-one cares.

    Example reports:

      https://bugzilla.redhat.com/show_bug.cgi?id=797687
      https://bugzilla.redhat.com/show_bug.cgi?id=867174
      https://bugzilla.redhat.com/show_bug.cgi?id=894865

    Signed-off-by: Dave Jones <davej@redhat.com>
    Cc: Andrew Morton <akpm@linux-foundation.org>
    Link: http://lkml.kernel.org/r/20130118175847.GA27662@redhat.com
    Signed-off-by: Ingo Molnar <mingo@kernel.org>

Some additional debug information:

Consider the following scenario.  A multi-processor system where there is
heavy IPI usage.

CPU 0
	- executes on_each_cpu() which is essentially a fancy wrapper
	around smp_call_function_many()
	- CPU 0 acquires it's per_cpu csd_data lock, and the
	call_function.queue's lock, call_function.lock, with irq
	protection.
	- CPU 0 adds this call function event to the the call_function.queue
	- CPU 0 releases the call_function.lock and allows irqs again

"Simultaneously" (of course keeping in mind the locks), the same thing occurs
on CPU 1.

CPU 0
	- smp_call_function_many() issues IPI

CPU 2
	- executes an IPI
	- generic_smp_call_function_interrupt() is called
	- goes through the call_function.queue and executes ALL queued
	functions.  This means the functions previously queued by CPU 0
	and CPU 1 are executed.
	- the cpumask for both functions are updated to remove CPU 2.

This same thing happens on all other CPUs in the system.  At the end of this
IPI (issued by CPU 0) the call function data's cpumask for both CPU 0 and
CPU 1 are EMPTY; both cpus's functions have executed in one IPI call.

CPU 0 merrily goes off and finishes its on_each_cpu ...

CPU 1, however, returns from the IPI and continues to execute in
smp_call_function_many() and does the following:

        /* Send a message to all CPUs in the map */
        arch_send_call_function_ipi_mask(data->cpumask);

which eventually resolves to

void default_send_IPI_mask_logical(const struct cpumask *cpumask, int vector)
{
        unsigned long mask = cpumask_bits(cpumask)[0];
        unsigned long flags;

        if (WARN_ONCE(!mask, "empty IPI mask"))
                return;

        local_irq_save(flags);
        WARN_ON(mask & ~cpumask_bits(cpu_online_mask)[0]);
        __default_send_IPI_dest_field(mask, vector, apic->dest_logical);
        local_irq_restore(flags);
}

and the WARN_ONCE is triggered because, as previously stated, the cpumask
is now empty.

A few things ...

1.  It seems that the smp_call functions did execute, so there appears to be no
worry there.

2.  There *might* be a concern due to ordering and waiting but I don't see one.

3.  Nothing is corrupted.

4. The code is basically the same upstream (we use llist and a few other things
are different ... but it looks like the issue is still there).  We are not
seeing any issues with this code path upstream and we're not seeing anything
other than the warning itself.  So ... I'm inclined just to do what upstream
did here and remove the WARN_ONCE (with an explanation of why it is okay).

4.  Is there still race here?  Yes.  In default_send_IPI_mask_logical() it
is entirely possible that the mask is set to zero AFTER the WARN_ONCE() by
an IPI on another processor.  But it seems like we've never seen that happen
in RHEL6 or upstream so I'm not worried about it.

I'm just going to remove the warning like they did upstream and be done with
it unless anyone can poke a hole in the above.

Cc: Dave Jones <davej@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>
Cc: Dave Anderson <anderson@redhat.com>
Cc: Paul Bunyan <pbunyan@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>